backend/requirements.txt
# Web framework
fastapi[standard]>=0.118.3

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Password hashing — bcrypt 5.x has Python 3.14 support
bcrypt>=5.0.0

# JWT — PyJWT replaces abandoned python-jose
PyJWT>=2.9.0

# Env + validation
python-dotenv>=1.0.0
pydantic[email]>=2.0.0

# Server
uvicorn[standard]>=0.30.0